# website/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WebEDT API",
    "version": "1.0.0",
    "description": "\nWebEDT is an AI-powered code editing platform. This API provides endpoints for:\n\n- **Authentication**: User registration, login, logout, and session management\n- **Sessions**: AI coding session management (create, list, update, delete)\n- **GitHub**: OAuth integration, repository operations, file management\n- **AI Execution**: Claude Remote Sessions for AI-assisted coding\n- **Store**: Game/asset store browsing and management\n- **Community**: Social features, channels, and real-time messaging\n\n## Authentication\n\nMost endpoints require authentication via session cookies. After logging in,\nthe server sets a session cookie that must be included in subsequent requests.\n\n## Response Format\n\nAll responses follow a standardized format:\n\n**Success Response:**\n```json\n{\n  \"success\": true,\n  \"data\": { ... },\n  \"timestamp\": \"2024-01-01T00:00:00.000Z\"\n}\n```\n\n**Error Response:**\n```json\n{\n  \"success\": false,\n  \"error\": {\n    \"message\": \"Error description\",\n    \"code\": \"ERROR_CODE\",\n    \"fields\": { \"fieldName\": [\"error message\"] }\n  },\n  \"timestamp\": \"2024-01-01T00:00:00.000Z\"\n}\n```\n\n## Rate Limiting\n\nAPI endpoints may be rate-limited at the infrastructure level.\n      ",
    "contact": {
      "name": "WebEDT Support",
      "email": "support@webedt.com"
    },
    "license": {
      "name": "Proprietary"
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_session",
        "description": "Session cookie set after successful login"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL_ERROR",
                  "BAD_REQUEST",
                  "SERVICE_UNAVAILABLE"
                ]
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "required": [
              "message"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "error",
          "timestamp"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "githubId": {
            "type": "string",
            "nullable": true
          },
          "githubAccessToken": {
            "type": "string",
            "nullable": true
          },
          "claudeAuth": {
            "type": "object",
            "nullable": true
          },
          "codexAuth": {
            "type": "object",
            "nullable": true
          },
          "geminiAuth": {
            "type": "object",
            "nullable": true
          },
          "preferredProvider": {
            "type": "string",
            "enum": [
              "claude",
              "codex",
              "gemini"
            ],
            "default": "claude"
          },
          "imageResizeMaxDimension": {
            "type": "integer",
            "nullable": true
          },
          "voiceCommandKeywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "defaultLandingPage": {
            "type": "string",
            "default": "store"
          },
          "preferredModel": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "message": "Unauthorized",
                "code": "UNAUTHORIZED"
              },
              "timestamp": "2024-01-01T00:00:00.000Z"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "message": "Validation failed",
                "code": "VALIDATION_ERROR",
                "fields": {
                  "email": [
                    "Invalid email address"
                  ]
                }
              },
              "timestamp": "2024-01-01T00:00:00.000Z"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "message": "Internal server error",
                "code": "INTERNAL_ERROR"
              },
              "timestamp": "2024-01-01T00:00:00.000Z"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "description": "Creates a new user account with email and password. Automatically logs in the user after successful registration.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "Password (minimum 8 characters)",
                    "example": "mySecurePassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "headers": {
              "Set-Cookie": {
                "description": "Session cookie for authentication",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "description": "Authenticates a user with email and password. Returns a session cookie for subsequent authenticated requests.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User's email address",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "User's password",
                    "example": "mySecurePassword123"
                  },
                  "rememberMe": {
                    "type": "boolean",
                    "description": "Extend session duration (90 days vs 30 days)",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "Set-Cookie": {
                "description": "Session cookie for authentication",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout user",
        "description": "Invalidates the current session and clears the session cookie.",
        "responses": {
          "200": {
            "description": "Logout successful",
            "headers": {
              "Set-Cookie": {
                "description": "Blank session cookie to clear authentication",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Logged out successfully"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/auth/session": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current session",
        "description": "Returns the current authenticated user's session information. Also refreshes OAuth tokens if needed.",
        "responses": {
          "200": {
            "description": "Session information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User"
                        },
                        "session": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "userId": {
                              "type": "string"
                            },
                            "expiresAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}